requirements.txt
SpeechRecognition
pyttsx3
pyaudio
requests
rich
google-generativeai